fastapi/pyproject.toml
[project]
name = "mfa-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.8",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
]

[dependency-groups]
dev = [
    "faker>=35.2.0",
]
